Linux port: MOVECODE is the 10th module to build (full) + cross-cutting roots
libbob_movecode.a = _MOVE.CPP (unity: Automove/Flyinit/Flymodel/Keyfly/Moveall/
Wpaction) + SAGMOVE.CPP, matching BOB.DSP. Key methodology insight: for modules
whose game build uses a _XXX.CPP unity TU, that unity IS the faithful build unit
— several fragment files (Flymodel/Moveall/Wpaction) have no own #includes and
only compile in unity context; compiling standalone TUs would miss them and
duplicate symbols.

Cross-cutting roots (no regression to the now-10 module libs):
- MYERROR.H assert/nassert under BOB_LINUX now expand to NOTHING (was ((void)0)).
  Bob uses asserts in statement positions ((void)0) can't satisfy: the
  no-trailing-semicolon idiom `assert(x) if(...)` and unbraced `if(c) assert(x);
  else ...`. Empty handles all of them.
- compat_types.h: LONG is now `long` (Win32 ABI), matching cstring.h's
  `typedef long LONG` — fixes the int32_t-vs-long conflicting-typedef.

Per-file fixes: AUTOMOVE (TimerCode.FRAMETIME->::, fpSqrt asm), SAGMOVE
(for-scope, animptr& rvalue->temp), FLYINIT/KEYFLY (raddef.h before radio.g for
USE_PHRASE_* aircraft enums; rchatter.h for _Radio/MESSAGE_STRUC; implicit-int
consts), FLYMODEL (mobileitem.ACList->::, static member).

Also: survey now filters to BOB.DSP game files (303 TUs) — excludes editor/tool
files (CEDITOR, MEGLOBAL, HARD320*, PERSONS5, vcl/devstudio-path files).

Ten module libs: math, lib3d, aircraft, general, files, input, missman, 3d, ai,
movecode.

+++ SRC/compat/compat_types.h
@@ -112,7 +112,7 @@ typedef unsigned short      WORD, *PWORD, *LPWORD;
 typedef short               SHORT, *PSHORT;
 typedef unsigned short      USHORT, *PUSHORT;
 typedef uint32_t            DWORD, *PDWORD, *LPDWORD;
-typedef int32_t             LONG, *PLONG, *LPLONG;
+typedef long int            LONG, *PLONG, *LPLONG;	/* Win32: LONG is long (matches cstring.h) */
 typedef uint32_t            ULONG, *PULONG;
 typedef int                 INT, *PINT, *LPINT;
 typedef unsigned int        UINT, *PUINT;
